ESTAGE-27 8fix erreur de calcul sur la durée du stage

+++ src/main/java/org/esup_portail/esup_stage/model/Convention.java
@@ -1141,7 +1141,7 @@ public class Convention extends ObjetMetier implements Exportable {
 
     public String getDureeExceptionnellePeriode() {
         if (this.getNbHeuresHebdo() != null) {
-            this.setDureeExceptionnellePeriode(PeriodeService.calculPeriodeOuvree(Float.parseFloat(this.getNbHeuresHebdo()), Integer.parseInt(this.getDureeExceptionnelle())));
+            this.setDureeExceptionnellePeriode(PeriodeService.calculPeriodeOuvree(Float.parseFloat(this.getNbHeuresHebdo()), Float.parseFloat(this.getDureeExceptionnelle())));
         }
         return dureeExceptionnellePeriode;
     }
